Add EWKT and EWKB encoders and decoders for SAP HANA.
This change adds EWKT and EWKB encoders and decoders for the SAP HANA
database. The existing Postgis EWKB and EWKT encoders don't write the
SRID if it's 0 or not set. HANA, however, always required the SRID to
be present in the EWKB or EWKT representation. This change adds encoders
and decoders for SAP HANA as extensions of the Postgis encoders which
always add the SRID to the output as well as the respective decoders
for symmetry reasons.