Add EWKT and EWKB encoders and decoders for SAP HANA.
This change adds EWKT and EWKB encoders and decoders for the SAP HANA
database. The existing Postgis EWKB and EWKT encoders don't write the
SRID if it's 0 or not set. HANA, however, always requires the SRID to
be present in the EWKB or EWKT representation. Another difference is how M and Z
coordinates are represented in the EWKB format. Postgis uses special flags in
the type code whereas HANA uses the type code ranges 1000+, 2000+, and 3000+ to
indicate these additional coordinates. This change adds encoders and decoders
for SAP HANA as extensions of the Postgis encoders which address those
differences.