Initial changes
1.
Added some ignores as we are working with eclipse

2.
apache-commons jar from destination is broken for us. Maybe the cause
is, that our sonartype nexus delivers already broken content.
Anyway: Fetched it from somewhere else

All of this changes should not be done in geolatte-geom master, as they
affect only GEOGRATs internal structure.