Encodes ORACLE Envelope and stores information in JTS.
This is a very special case and maybe only needed in GEOGRAT GISX.

If a geometry can not be stored in ORACLE Spatial, GEOGRAT stores it as
RECTANGLE, in order to fetch it with spatial queries.
If at querying such a rectangle is found, the geometry can be fetched
from different storage.

To achieve this in ORACLE spatial, I decided to store this information
in user data of JTS geometry.
To achieve this, it is first converted to a PolygonEnvelope (subclass of
Polygon, so there should be no problem in rest of code).
When converting to JTS, a constant is set in user data.